Rename ovsdb_users to ovsdb-client
Part of the changes to convert the ovsdb_users group to the ovsdb-client
group.  Note that test_openvswitch_ct_socket_permission.py checks both
groups so that dpendent changes in the ops_build repo can pass the
gate tests.  Once that change is in, the reference to ovsdb_users will
be removed.

Signed-off-by: Scott Gulland <scott.gulland@hpe.com>