Change track: retain column values of deleted rows
When change tracking is enabled, only the row uuid of deleted rows is
available. The column values are unparsed and the ovsdb_datum values
are cleared so they are not available for inspection. This change
leaves the column values around for inspection. The column values are
cleared (unparsed) upon clearing the change track.

Signed-off-by: Shad Ansari <shad.ansari@hpe.com>
Signed-off-by: Sebastian Arguello <sebastian.arguello@hpe.com>