Enable OpenFlow
Uncommented code in OVS to enable the OpenFlow agent. The OpenFlow
agent runs in the ops-switchd process. It talks to the remote
OpenFlow Controller and installs OpenFlow forwarding rules in the
ASIC. The ASIC driver for Broadcom silicon is provided by the
changes done in the opennsl-plugin.

Also made changes to switch to the Management namespace while
creating a socket to be able to connect to the OpenFlow
Controller via the service port.

Change-Id: Ibc86f8645a2088bf074d3ab4367e25cbfe2d866c
Signed-off-by: Amitabha Sen <amitabha.sen@broadcom.com>