new: dev: Ofproto mirror bundle.
The ofproto_mirror_bundle struct is to enable mirror_configure to pair a
mirror source or destination port with whatever bridge or VRF ofproto it is
currently associated with.
This association of ofprotos with ports allows the PD layer to locate a given
port via it's ofproto number when the mirror is created/modified via mirror_set.

Change-Id: Ibe984dffd8c271d7bd771a547e68c433222abb3c
Signed-off-by: midhan <nitish.midha@hpe.com>