new: dev: Enable OVS tests inside OPS
This patch enables must of the tests from the OVS repo to be run inside
the OPS repo.
Two scripts (run-ovs-tests.sh and ovs-tests.sh) are added that make the
changes needed.
To run the tests just need to execute run-ovs-tests.sh from the
ops-openvswitch folder.
Review the comments inside the scripts to get more information.

Change-Id: Iac1fd909529d1a4d16655a5a19d7cbed31ef18b6
Signed-off-by: arnoldo.lutz.guevara@hpe.com <arnoldo.lutz.guevara@hpe.com>